chore: refactor replay register fn to misc utils
